coding: PRされたものの調整中

+++ cogs/server-tool/globalchat.py
@@ -3,6 +3,8 @@
 from typing import TypedDict, NamedTuple
 from collections.abc import AsyncIterator
 
+from collections import defaultdict
+
 import discord
 from discord.ext import commands
 
@@ -11,13 +13,14 @@ from core import Cog, RT, t, DatabaseManager, cursor
 from rtutil.utils import webhook_send
 from rtlib.common.json import dumps, loads
 
+from data import FORBIDDEN
+
 from .__init__ import FSPARENT
 
 
 class Setting(TypedDict, total=False):
     "グローバルチャットの設定"
     password: str | None
-
 class Data(NamedTuple):
     name: str
     author_id: int
@@ -27,13 +30,17 @@ class Data(NamedTuple):
 class DataManager(DatabaseManager):
     "セーブデータを管理します"
 
+    MAX_GLOBAL_CHAT_COUNT = 3
+    MAX_CHANNEL_COUNT = 30
+
     def __init__(self, bot: RT):
         self.pool = bot.pool
         self.bot = bot
+        self.caches: defaultdict[str, list[int]] = defaultdict(list)
 
     async def prepare_table(self) -> None:
         await cursor.execute(
-            """CREATE TABLE IF NOT EXISTS GlobalChat(
+            """CREATE TABLE IF NOT EXISTS GlobalChat (
                 Name TEXT, AuthorId BIGINT, Setting JSON
             );"""
         )
@@ -48,37 +55,63 @@ class DataManager(DatabaseManager):
                 Name TEXT, ChannelId BIGINT
             );"""
         )
+        # キャッシュを用意する。
+        async for row in self.fetchstep(cursor, "SELECT * FROM GlobalChatChannel;"):
+            self.caches[row[0]].append(row[1])
 
-    async def connect(self, name: str, channel_id: int) -> None:
-        "グローバルチャットに接続します。"
+    async def is_exists_with_error(self, name: str, **_) -> None:
+        "`.is_exists`を実行して、既に存在する場合のみエラーを発生させます。"
+        if await self.is_exists(name, cursor=cursor):
+            raise Cog.BadRequest({
+                "en": "This globalchat already exists.",
+                "ja": "このグローバルチャットは既に存在しています。"
+            })
+
+    async def is_connected_with_error(self, channel_id: int, **_) -> None:
+        "`.is_connected`を実行して、既に接続している場合はエラーを発生させます。"
         if await self.is_connected(channel_id, cursor=cursor):
             raise Cog.BadRequest({
                 "en": "You are already connected to this globalchat.",
                 "ja": "既にこのグローバルチャットに接続しています。"
             })
+
+    async def before_common_check(self, name: str, channel_id: int, **_) -> None:
+        "`.is_exists_with_error`と`.is_connected_with_error`を実行します。"
+        await self.is_exists_with_error(name, cursor=cursor)
+        await self.is_connected_with_error(channel_id, cursor=cursor)
+
+    async def connect(self, name: str, channel_id: int) -> None:
+        "グローバルチャットに接続します。"
+        await self.before_common_check(name, channel_id)
         await cursor.execute(
             "INSERT INTO GlobalChatChannel VALUES (%s, %s);",
             (name, channel_id)
         )
+        self.caches[name].append(channel_id)
 
     async def create(
-        self, name: str, author_id: int, channel_id: int, setting: Setting
+        self, name: str, author_id: int,
+        channel_id: int, setting: Setting
     ) -> None:
         "主にグローバルチャットを作るために使います。"
-        if await self.is_connected(channel_id, cursor=cursor):
-            raise Cog.BadRequest({
-                "en": "You are already connected to globalchat.",
-                "ja": "既にグローバルチャットに接続しています。"
-            })
+        await self.before_common_check(name, channel_id)
+        # グローバルチャットを作成した人が、最大作成可能個数分既にグローバルチャットを作っているのなら、エラーを起こす。
         await cursor.execute(
-            "SELECT * FROM GlobalChat WHERE Name = %s;",
-            (name,)
+            "SELECT COUNT(AuthorId) FROM GlobalChat WHERE AuthorId = %s;",
+            (author_id,)
         )
-        if await cursor.fetchone() is not None:
+        if (await cursor.fetchone())[0] > self.MAX_GLOBAL_CHAT_COUNT:
             raise Cog.BadRequest({
-                "en": "This globalchat already exists.",
-                "ja": "このグローバルチャットは既に存在しています。"
+                "ja": "あなたはグローバルチャットを作りすぎです。\nそのためグローバルチャットを作ることができません。",
+                "en": "You are creating too much global chat.\nYou cannot create a global chat because of that."
             })
+        # 接続数が最大接続数になっている場合は拒否する。
+        if len(self.caches[name]) >= 30:
+            raise Cog.BadRequest({
+                "ja": "これ以上このグローバルチャットにチャンネルを接続することができません。",
+                "en": "No more channels can be connected for this global chat."
+            })
+        # グローバルチャットの作成を行う。
         await cursor.execute(
             "INSERT INTO GlobalChat VALUES (%s, %s, %s);",
             (name, author_id, dumps(setting))
@@ -87,16 +120,13 @@ class DataManager(DatabaseManager):
             "INSERT INTO GlobalChatChannel VALUES (%s, %s);",
             (name, channel_id)
         )
+        self.caches[name].append(channel_id)
 
     async def is_connected(self, channel_id: int, **_) -> bool:
         "これはすでに接続されているか確認するものです。"
-        await cursor.execute(
-            "SELECT * FROM GlobalChatChannel WHERE ChannelId = %s LIMIT 1;",
-            (channel_id,)
-        )
-        return bool(await cursor.fetchone())
+        return (await self.get_name(channel_id, cursor=cursor)) is not None
 
-    async def is_existed(self, name: str) -> bool:
+    async def is_exists(self, name: str, **_) -> bool:
         "すでにグローバルチャットが存在するか確認します。"
         await cursor.execute(
             "SELECT * FROM GlobalChat WHERE Name = %s LIMIT 1;",
@@ -113,17 +143,12 @@ class DataManager(DatabaseManager):
         if row := await cursor.fetchone():
             return loads(row[0])["password"] == password
 
-    async def get_channels(self, name: str, **_) -> AsyncIterator[discord.TextChannel]:
+    async def get_channel_ids(self, name: str, **_) -> AsyncIterator[int]:
         "グローバルチャットに接続しているチャンネルを名前使って全部取得します。"
         async for (channel_id,) in self.fetchstep(
             cursor, "SELECT ChannelId FROM GlobalChatChannel WHERE Name = %s;", (name,)
         ):
-            channel = self.bot.get_channel(channel_id)
-            if isinstance(channel, discord.TextChannel):
-                yield channel
-            else:
-                await self.disconnect(channel_id, cursor=cursor)
-                self.bot.print(f"[GlobalChat] Delete: {channel_id}")
+            yield channel_id
 
     async def get_name(self, channel_id: int, **_) -> str | None:
         "チャンネルから接続しているグローバルチャット名を取得します。"
@@ -135,22 +160,26 @@ class DataManager(DatabaseManager):
             return row[0]
 
     async def disconnect(self, channel_id: int, **_) -> None:
-        "グローバルチャットから接続をやめます"
+        "グローバルチャットから接続をやめます。"
+        if (name := await self.get_name(channel_id, cursor=cursor)) is None:
+            raise Cog.BadRequest({
+                "ja": "そのチャンネルは接続されていません。",
+                "en": "That channel is not connected."
+            })
         await cursor.execute(
             "DELETE FROM GlobalChatChannel WHERE ChannelId = %s;",
             (channel_id,)
         )
+        self.caches[name].remove(channel_id)
 
-    async def insert_message(
-        self, source: int, channel_id: int, message_id: int
-    ) -> None:
+    async def insert_message(self, source: int, channel_id: int, message_id: int) -> None:
         "メッセージを保存します。"
         await cursor.execute(
             "INSERT INTO GlobalChatMessage VALUES (%s, %s, %s);",
             (source, channel_id, message_id)
         )
 
-    async def fetch_global_chat(self, name: str) -> Data:
+    async def get_setting(self, name: str) -> Data:
         "グローバルチャットを取得します。"
         await cursor.execute(
             "SELECT * FROM GlobalChat WHERE Name = %s LIMIT 1;",
@@ -158,44 +187,54 @@ class DataManager(DatabaseManager):
         )
         if row := await cursor.fetchone():
             return Data(*row)
-        else:
-            raise Cog.BadRequest({
-                "en": "That globalchat does not exist.",
-                "ja": "そのグローバルチャットは存在しません。"
-            })
+        raise Cog.BadRequest({
+            "en": "That globalchat does not exist.",
+            "ja": "そのグローバルチャットは存在しません。"
+        })
+
+
+class GlobalChatEventContext(Cog.EventContext):
+    "グローバルチャットに送信されたメッセージを、チャンネルに送信する際に発生するイベントのコンテキストです。"
+
+    channel: discord.TextChannel
+    message: discord.Message
 
 
 class GlobalChat(Cog):
     "グローバルチャットのコグです。"
 
-    WEBHOOK_NAME = "rt-globalchat-webhook"
-
     def __init__(self, bot: RT):
         self.bot = bot
+        self.pool = self.bot.pool
         self.data = DataManager(bot)
 
     async def cog_load(self):
         await self.data.prepare_table()
 
     @commands.group(
-        description="The command of globalchat.",
-        aliases=("gc", "gchat"), fsparent=FSPARENT
+        description="The command of globalchat.", fsparent=FSPARENT,
+        aliases=("gc", "gchat", "グローバルチャット", "ぐろちゃ")
     )
     async def globalchat(self, ctx):
         await self.group_index(ctx)
+
+    async def check_text_channel(self, ctx: commands.Context) -> None:
+        "テキストチャンネル以外のコンテキストの場合はエラーが発生します。"
+        if not isinstance(ctx.channel, discord.TextChannel):
+            raise Cog.BadRequest({
+                "ja": "グローバルチャットに接続させるチャンネルはテキストチャンネルでなければいけません。",
+                "en": "The channel to be connected to global chat must be a text channel."
+            })
 
     @globalchat.command(description="Create globalchat.", aliases=("make", "add", "作成")
     )
     @discord.app_commands.describe(name="Global chat name", password="Password")
     async def create(self, ctx, name: str, *, password: str | None = None):
-        await self.data.create(
-            name, ctx.author.id, ctx.channel.id, {
-                "password": password
-            }
-        )
-        await ctx.reply(t(dict(
-            en="Created", ja="作成しました。"
-        ), ctx))
+        await self.check_text_channel(ctx)
+        await self.data.create(name, ctx.author.id, ctx.channel.id, {
+            "password": password
+        })
+        await ctx.reply(t(dict(en="Created", ja="作成しました。"), ctx))
 
     @globalchat.command(
         description="Connect to global chat.",
@@ -203,18 +242,14 @@ class GlobalChat(Cog):
     )
     @discord.app_commands.describe(name="Global chat name", password="Password")
     async def connect(self, ctx, name: str, *, password: str | None = None):
-        if not await self.data.is_existed(name):
-            return await ctx.reply(t(dict(
-                en="Not found", ja="見つかりませんでした。"
-            ), ctx))
+        await self.check_text_channel(ctx)
         if not await self.data.check_password(name, password):
             return await ctx.reply(t(dict(
-                en="Wrong password", ja="パスワードが間違っています。"
+                en="Wrong password",
+                ja="パスワードが間違っています。"
             ), ctx))
         await self.data.connect(name, ctx.channel.id)
-        await ctx.reply(t(dict(
-            en="Connected", ja="接続しました。"
-        ), ctx))
+        await ctx.reply(t(dict(en="Connected", ja="接続しました。"), ctx))
 
     @globalchat.command(
         description="Disconnect from globalchat.",
@@ -223,22 +258,23 @@ class GlobalChat(Cog):
     async def disconnect(self, ctx):
         await self.data.disconnect(ctx.channel.id)
         await ctx.reply(t(dict(
-            ja="グローバルチャットから退出しました", en="Leave from globalchat"
+            ja="グローバルチャットから退出しました",
+            en="Leave from globalchat"
         ), ctx))
 
     _help = Cog.HelpCommand(globalchat).merge_description("headline", ja="グローバルチャット関連です。")
     _help.add_sub(Cog.HelpCommand(create)
-                  .merge_description("headline", ja="グローバルチャットを作成します。")
-                  .add_arg("name", "str", "Optional",
-                           ja="グローバルチャット名", en="Globalchat name")
-                  .add_arg("password", "str", "Optional",
-                           ja="パスワード", en="Password"))
+        .merge_description("headline", ja="グローバルチャットを作成します。")
+        .add_arg("name", "str", "Optional",
+            ja="グローバルチャット名", en="Globalchat name")
+        .add_arg("password", "str", "Optional",
+            ja="パスワード", en="Password"))
     _help.add_sub(Cog.HelpCommand(connect)
-                  .merge_description("headline", ja="グローバルチャットに接続します。")
-                  .add_arg("name", "str", "Optional",
-                           ja="グローバルチャット名", en="GlobalChat name")
-                  .add_arg("password", "str", "Optional",
-                           ja="パスワード", en="Password"))
+        .merge_description("headline", ja="グローバルチャットに接続します。")
+        .add_arg("name", "str", "Optional",
+            ja="グローバルチャット名", en="GlobalChat name")
+        .add_arg("password", "str", "Optional",
+            ja="パスワード", en="Password"))
     _help.add_sub(Cog.HelpCommand(disconnect).merge_description(
         "headline", ja="グローバルチャットから退出します。"))
     del _help
@@ -250,18 +286,43 @@ class GlobalChat(Cog):
 
         async with self.bot.pool.acquire() as connection:
             async with connection.cursor() as cursor:
-                if not await self.data.is_connected(message.channel.id, cursor=cursor) or \
-                     await self.data.get_name(message.channel.id, cursor=cursor) is None:
+                # 接続しているかの確認と名前の取得を行う。
+                if not await self.data.is_connected(message.channel.id, cursor=cursor) or (
+                    name := await self.data.get_name(message.channel.id, cursor=cursor)
+                ) is None:
                     return
-                async for channel in self.data.get_channels(
-                    await self.data.get_name(message.channel.id, cursor=cursor), cursor=cursor
-                ):
-                    if message.channel.id != channel.id:
+                # メッセージの送信を行う。
+                async for channel_id in self.data.get_channel_ids(name, cursor=cursor):
+                    if message.channel.id == channel_id:
+                        continue
+                    # チャンネルの取得を行う。
+                    channel = await self.bot.search_channel(channel_id)
+                    assert isinstance(channel, discord.TextChannel)
+                    if channel is None:
+                        await self.data.disconnect(channel_id, cursor=cursor)
+                        continue
+                    # 送信を行う。
+                    error = None
+                    try:
                         await webhook_send(
-                            channel, message.author,
-                            message.clean_content,
-                            files=[await attachment.to_file() for attachment in message.attachments]
+                            channel, message.author, message.clean_content, files=[
+                                await attachment.to_file()
+                                for attachment in message.attachments
+                            ]
                         )
+                    except discord.Forbidden:
+                        error = FORBIDDEN
+                    except Exception:
+                        ...
+                    self.bot.rtevent.dispatch("on_global_chat_message", GlobalChatEventContext(
+                        self.bot, channel.guild, error, {
+                            "ja": "グローバルチャットからのメッセージの襲来",
+                            "en": "An assault of messages from global chat"
+                        }, self.text_format({
+                            "ja": "送信対象：{name}", "en": "Target: {name}"
+                        }, name=self.name_and_id(channel)), self.globalchat, error,
+                        channel=channel, message=message
+                    ))
 
 
 async def setup(bot: RT) -> None:
